Add damage numbers

+++ src/sys/camera_system.cpp
@@ -20,6 +20,6 @@ void CameraSystem::update(World &w)
 							  .pos;
 
 		cf_camera_dimensions(camera_size.x, camera_size.y);
-		cf_camera_look_at(player_pos.x, player_pos.y);
+		cf_camera_look_at((int)player_pos.x, (int)player_pos.y);
 	}
 }
